+++ src/plugins/process/CMakeLists.txt
@@ -24,4 +24,5 @@ add_subdirectory(tls)
 if (ENABLE_PROCESS_EXPERIMENTAL)
 	add_subdirectory(sip)
 	add_subdirectory(rtsp)
+	add_subdirectory(mpls)
 endif()
